gptp SYSTIM issue change (#14)
When the PTP daemon is running it takes control over SYSTIM (using kernel's clock_adjtime with clockid from /dev/ptpX). At this point any change made to SYSTIM by the driver can confuse the daemon and in the result break some use cases. Previously SYSTIM was changed every time the igb_reset was called (in the igb_ptp_reset). The solution is to initialize SYSTIM once in the driver and don't touch it (change it) again.