kmod: Use fallthrough keyword to avoid implicit fallthoughs
In kcompat.h, we added the 'fallthrough' pseudo keyword introduced
in kernel mainline commit (294f69e662d15). We use this pseudo keyword
to explicitly fallthrough and avoid the implicit-fallthrough warnings.

Since Kernel version >= 5.17 in x86 defconfig CONFIG_WERROR=y
is set. So we need to get rid of those warnings to smoothly compile
for kernels using this config

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>